requirements.txt
alembic==1.11.1
SQLAlchemy==2.0.15
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
asyncpg==0.27.0
geopy==2.3.0
rocketry==2.5.1
fastapi-filter==0.6.0